refactor: 자유게시판 폼(작성, 수정) 컴포넌트 작업
